move files out of lib subdir
