+++ .eslintrc.yml
@@ -1,7 +1,10 @@
 env:
-    browser: true
-    commonjs: true
-    es6: true
+  browser: true
+  # commonjs: true
+  node: true
+  es6: true
+parserOptions:
+  sourceType: module
 extends: 
   - 'eslint:recommended'
   - 'plugin:react/recommended'
