+++ package.json
@@ -1,9 +1,12 @@
 {
     "name": "@kbase/ui-lib",
-    "version": "0.3.6",
+    "version": "0.4.0",
     "description": "A typescript library to support kbase-ui and plugins",
     "main": "lib/index.js",
     "types": "lib/index.d.ts",
+    "directories": {
+        "lib": "lib"
+    },
     "files": [
         "/lib/**/*"
     ],
@@ -12,8 +15,8 @@
         "copy-assets": "copyfiles -u 1  \"./src/**/*.css\" \"./lib/\"",
         "clean": "rm -rf ./lib/*",
         "build": "bash scripts/build.bash",
-        "prepare": "bash scripts/prepare.bash",
-        "prepublish": "jest src",
+        "prepack": "bash scripts/build.bash",
+        "prepublishOnly": "jest src",
         "start": "tsc --watch",
         "generate-api-docs": "typedoc --out api-docs src"
     },
