+++ scripts/build.bash
@@ -1,8 +1,6 @@
-tsc
 rm -rf ./lib
-rm -rf ./templib
-copyfiles -u 1  "./build/**/*" "./templib/"
-copyfiles -u 1  "./src/**/*.css" "./templib/"
-copyfiles -u 1 "./src/style/**/*" "./templib/"
-copyfiles -u 1 "./src/custom/**/*" "./templib/"
-mv ./templib ./lib
+tsc
+copyfiles -u 1  "./build/**/*" "./lib/"
+copyfiles -u 1  "./src/**/*.css" "./lib/"
+copyfiles -u 1 "./src/style/**/*" "./lib/"
+copyfiles -u 1 "./src/custom/**/*" "./lib/"
