+++ src/comm/dynamicServices/formatsx/sesar/lists/navigation-type.tsv
