+++ src/comm/dynamicServices/formatsx/sesar/lists/platform-type.tsv
