+++ src/simpleObject.ts
